requirements.txt
pillow
imagehash
numpy
tqdm